+++ test/generateDefinition.test.ts
@@ -1,20 +1,18 @@
 import {describe, expect, test} from "vitest";
-import type {Models as ZHModels} from "zigbee-herdsman";
 import {Zcl} from "zigbee-herdsman";
 import {findByDevice, generateExternalDefinitionSource} from "../src";
 import * as fz from "../src/converters/fromZigbee";
 import {repInterval} from "../src/lib/constants";
-import type {Definition} from "../src/lib/types";
 import {type AssertDefinitionArgs, assertDefinition, mockDevice, reportingItem} from "./utils";
 
 const assertGeneratedDefinition = async (args: AssertDefinitionArgs & {externalDefinitionSource?: string}) => {
-    const getDefinition = async (device: ZHModels.Device): Promise<Definition> => findByDevice(device, true);
-    const definition = await getDefinition(args.device);
+    const definition = await findByDevice(args.device, true);
     expect(definition.model).toEqual(args.device.modelID);
     if (args.externalDefinitionSource) {
         expect((await generateExternalDefinitionSource(args.device)).trim()).toEqual(args.externalDefinitionSource.trim());
     }
-    return await assertDefinition({findByDeviceFn: getDefinition, ...args});
+    // prevent unnecessary duplicate call to `findByDevice` by passing already retrieved definition
+    return await assertDefinition({findByDeviceFn: () => Promise.resolve(definition), ...args});
 };
 
 describe("GenerateDefinition", () => {
@@ -532,6 +530,102 @@ export default {
     extend: [m.electricityMeter({"cluster":"electrical","electricalMeasurementType":"dc","endpointNames":["2"]})],
 };
             `,
+        });
+    });
+
+    test("input(genBinaryInput), output(genBinaryOutput, genAnalogOutput)", async () => {
+        const attr10 = {
+            genBinaryInput: {
+                description: "my_binary_name",
+            },
+            genAnalogOutput: {
+                description: "my_output_name",
+                applicationType: 0,
+                engineeringUnits: 62,
+                minPresentValue: 0.0,
+                maxPresentValue: 30.0,
+                resolution: 0.1,
+                presentValue: 15.0,
+            },
+        };
+
+        await assertGeneratedDefinition({
+            device: mockDevice({
+                modelID: "temp",
+                endpoints: [
+                    {ID: 10, inputClusters: ["genBinaryInput", "genBinaryOutput", "genAnalogOutput"], outputClusters: [], attributes: attr10},
+                ],
+            }),
+            meta: undefined,
+            fromZigbee: [
+                expect.objectContaining({cluster: "genBinaryInput"}),
+                expect.objectContaining({cluster: "genBinaryOutput"}),
+                expect.objectContaining({cluster: "genAnalogOutput"}),
+            ],
+            toZigbee: ["my_binary_name", "binary_output_10", "my_output_name"],
+            exposes: ["binary_output_10", "my_binary_name", "my_output_name"],
+            bind: {10: ["genBinaryInput", "genBinaryOutput", "genAnalogOutput"]},
+            read: {
+                10: [
+                    ["genBinaryOutput", ["description"], {sendPolicy: "immediate", disableRecovery: true}],
+                    ["genBinaryInput", ["presentValue"]],
+                    ["genBinaryOutput", ["presentValue"]],
+                    ["genAnalogOutput", ["presentValue"]],
+                ],
+            },
+            configureReporting: {
+                10: [
+                    ["genBinaryInput", [reportingItem("presentValue", 0, 65000, 1)]],
+                    ["genBinaryOutput", [reportingItem("presentValue", 0, 65000, 1)]],
+                    ["genAnalogOutput", [reportingItem("presentValue", 0, 65000, 1)]],
+                ],
+            },
+        });
+    });
+
+    test("input(genAnalogInput), x2 endpoints", async () => {
+        const attr10 = {
+            genAnalogInput: {
+                description: "my_custom_name",
+                applicationType: 0,
+                engineeringUnits: 62,
+                minPresentValue: 0.0,
+                maxPresentValue: 30.0,
+                resolution: 0.1,
+                presentValue: 15.0,
+            },
+        };
+
+        await assertGeneratedDefinition({
+            device: mockDevice({
+                modelID: "temp",
+                endpoints: [
+                    {ID: 10, inputClusters: ["genAnalogInput"], outputClusters: [], attributes: attr10},
+                    {ID: 11, inputClusters: ["genAnalogInput"], outputClusters: []},
+                ],
+            }),
+            meta: {multiEndpoint: true},
+            fromZigbee: [expect.objectContaining({cluster: "genAnalogInput"}), expect.objectContaining({cluster: "genAnalogInput"})],
+            toZigbee: ["my_custom_name", "analog_input"],
+            exposes: ["analog_input_11", "my_custom_name_10"],
+            bind: {10: ["genAnalogInput"], 11: ["genAnalogInput"]},
+            read: {
+                10: [["genAnalogInput", ["presentValue"]]],
+                11: [
+                    ["genAnalogInput", ["description"], {disableRecovery: true, sendPolicy: "immediate"}],
+                    ["genAnalogInput", ["applicationType"], {disableRecovery: true, sendPolicy: "immediate"}],
+                    ["genAnalogInput", ["engineeringUnits"], {disableRecovery: true, sendPolicy: "immediate"}],
+                    ["genAnalogInput", ["minPresentValue"], {disableRecovery: true, sendPolicy: "immediate"}],
+                    ["genAnalogInput", ["maxPresentValue"], {disableRecovery: true, sendPolicy: "immediate"}],
+                    ["genAnalogInput", ["resolution"], {disableRecovery: true, sendPolicy: "immediate"}],
+                    ["genAnalogInput", ["presentValue"]],
+                ],
+            },
+            configureReporting: {
+                10: [["genAnalogInput", [reportingItem("presentValue", 0, 65000, 1)]]],
+                11: [["genAnalogInput", [reportingItem("presentValue", 0, 65000, 1)]]],
+            },
+            endpoints: {"10": 10, "11": 11},
         });
     });
 });